代码/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
networkx>=3.0
matplotlib>=3.7.0
pandas>=2.0.0
pyyaml>=6.0